+++ src/Expressif.LanguageServer.Core/Evaluation/IExpressionEvaluationService.cs
@@ -2,5 +2,5 @@ namespace Expressif.LanguageServer.Core.Evaluation;
 
 public interface IExpressionEvaluationService
 {
-    ExpressionEvaluationResult Evaluate(string expression, string input);
+    ExpressionEvaluationResult Evaluate(string expression, string? input = null);
 }